Created ABC for DataCatalog
